added documentation

+++ Java/src/test/resources/sql-schema.sql
@@ -13,7 +13,7 @@ CREATE TABLE IF NOT EXISTS `customers` (
 CREATE TABLE IF NOT EXISTS `items` (
     `id` INT(11) NOT NULL AUTO_INCREMENT,
     `item_name` VARCHAR(50) DEFAULT NULL,
-    `item_value` INT(11) DEFAULT NULL,
+    `item_value` DOUBLE DEFAULT NULL,
     PRIMARY KEY (`id`)
 );
 
@@ -31,7 +31,7 @@ CREATE TABLE IF NOT EXISTS `orders`(
 
 `order_id` INT(11) NOT NULL AUTO_INCREMENT,
 `customer_id` INT(11) NOT NULL,
-`total_order_cost` FLOAT(11) NOT NULL,
+`total_order_cost` DOUBLE NOT NULL,
 PRIMARY KEY (order_id)
 
 );
